moved git-related stuff to a separate module, introduced compatibilities
It so happens that the git team often changes bahvior of certain git
applets between releases and we have to support different versions of
git. Debian, for example has 1.4.x.y versions of git-core in stable
and testing (as of today) and 1.5.x.y in unstable. This little fact
inspired the 'compatibility' functionality which detects version of
installed git (in run-time) and selects relevant compatibility flags.

Since the amount of code that is tightly working with git tools is
growing with time, all this code was separated-out to a new module;
automake thingy regenerated.

Several comments added to make the code look more like commented.
Signed-off-by: Alexander Shishkin <virtuoso@slind.org>